Okay, now I want to make a MAJOR UI/UX overhaul!
The initial page is very clustered because of the huge amount of information in the initial screen.

The proposal: I want to keep only the extremely important information on the initial screen, and make a UI pages or tabs for a more detailed view of quests, lorebook entries, characters stats/inventory.