Okay this has fixed it, however I've noticed something strange with the NPCs relationships, because at the start of the series, when Subaru has been transported to this other world, Subaru hasn't met Emilia nor Puck, so why is their relationship with him above than 0? Also, why does subaru have a Rusty Iron Shortsword, didn't he just get transported?
Another thing:
GAME-MASTER's avatar
GAME-MASTER(admin)

The blinding light faded, replaced by the bustling sounds of a vibrant street. Market stalls line cobblestone paths, filled with exotic goods and strange, non-human vendors. The air hums with unfamiliar languages and the scent of unfamiliar spices. People in fantastical attire move purposefully, some with animalistic features. It's a stark contrast to the sterile convenience store you were just in.

This initial message doesn't give a little push to the user for him to pursue the start of the act.