Okay so I want to propose an awesome new idea!:
Activation  points for the start of arcs!

It would nice if instead of completing the main objectives to complete an arc the user would need to achieve a certain "story status" I don't know how phrase it.

And when the arc starts/ends there would be a cool notification with an animation on the chat or something like that. It would also be nice if there was a success metric for the arc, for example, Subaru has managed to save Emilia, but killed many people while trying to save her, so his success score would be 10/100, do you understand?

What do you think? The main quests should still be a thing, but the completion of ALL of them shouldn't determine the start/end of an arc.